+++ src/Tests.WebFake/CRUDResponses/DeleteResponseEntry.cs
@@ -2,7 +2,7 @@
 using FatCat.Fakes;
 using FatCat.Toolkit.WebServer.Testing;
 using FatCat.WebFake.Endpoints;
-using FatCat.WebFake.Models;
+using FatCat.WebFakeApi.Models;
 using Xunit;
 
 namespace Tests.FatCat.WebFake.CRUDResponses;
