+++ src/Tests.WebFake/CRUDResponses/GetAllClientRequests.cs
@@ -1,8 +1,8 @@
 ﻿using FakeItEasy;
 using FatCat.Fakes;
 using FatCat.Toolkit.WebServer.Testing;
+using FatCat.WebFake;
 using FatCat.WebFake.Endpoints;
-using FatCat.WebFake.Models;
 using Xunit;
 
 namespace Tests.FatCat.WebFake.CRUDResponses;
