+++ src/Tests.WebFake/VerbTests/DeleteTests.cs
@@ -1,6 +1,6 @@
 ﻿using FatCat.Toolkit.WebServer;
 using FatCat.WebFake.Endpoints;
-using FatCat.WebFake.Models;
+using FatCat.WebFakeApi.Models;
 
 namespace Tests.FatCat.WebFake.VerbTests;
