+++ src/Tests.WebFake/VerbTests/VerbTests.cs
@@ -3,8 +3,9 @@ using FakeItEasy;
 using FatCat.Fakes;
 using FatCat.Toolkit.WebServer;
 using FatCat.Toolkit.WebServer.Testing;
+using FatCat.WebFake;
 using FatCat.WebFake.Endpoints;
-using FatCat.WebFake.Models;
+using FatCat.WebFakeApi.Models;
 using FluentAssertions;
 using Humanizer;
 using Newtonsoft.Json;
