+++ src/WebFake/ClientRequestCacheItem.cs
@@ -1,7 +1,8 @@
 ﻿using FatCat.Toolkit;
 using FatCat.Toolkit.Caching;
+using FatCat.WebFakeApi.Models;
 
-namespace FatCat.WebFake.Models;
+namespace FatCat.WebFake;
 
 public class ClientRequestCacheItem : EqualObject, ICacheItem
 {
