+++ src/WebFake/Endpoints/PostEndpoint.cs
@@ -2,7 +2,7 @@
 using FatCat.Toolkit.Caching;
 using FatCat.Toolkit.Threading;
 using FatCat.Toolkit.WebServer;
-using FatCat.WebFake.Models;
+using FatCat.WebFakeApi.Models;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
