+++ src/WebFake/ResponseCacheItem.cs
@@ -1,7 +1,8 @@
 ﻿using FatCat.Toolkit;
 using FatCat.Toolkit.Caching;
+using FatCat.WebFakeApi.Models;
 
-namespace FatCat.WebFake.Models;
+namespace FatCat.WebFake;
 
 public class ResponseCacheItem : EqualObject, ICacheItem
 {
