--- src/WebFakeApi/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace WebFakeApi;
-
-public class Class1
-{
-
-}
