+++ src/WebFakeApi/Models/ClientRequest.cs
@@ -2,7 +2,7 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace FatCat.WebFake.Models;
+namespace FatCat.WebFakeApi.Models;
 
 public class ClientRequest : EqualObject
 {
